chore: added snapshot & e2e specs
